# docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Artifex API",
        "description": "二维游戏美术协作与 AI 资产生成平台 API",
        "version": "1.4.0",
        "contact": { "name": "Artifex Team" }
    },
    "servers": [
        { "url": "http://127.0.0.1:3000", "description": "本地开发" }
    ],
...
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": ["auth"],
                "summary": "注册",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["username", "email", "password"],
                                "properties": {
                                    "username": { "type": "string" },
                                    "email": { "type": "string", "format": "email" },
                                    "password": { "type": "string", "minLength": 6 }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "注册成功" },
                    "400": { "description": "参数错误" },
                    "409": { "description": "邮箱/用户名已存在" }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": ["auth"],
                "summary": "登录",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["email", "password"],
                                "properties": {
                                    "email": { "type": "string" },
                                    "password": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "登录成功" },
                    "401": { "description": "邮箱或密码错误" }
                }
            }
        },
...
        "/api/me/profile": {
            "put": {
                "tags": ["auth"],
                "summary": "更新用户资料",
                "requestBody": {
                    "content": { "application/json": { "schema": { "type": "object", "properties": { "profile": { "type": "object" } } } } }
                },
                "responses": { "200": { "description": "更新成功" } }
            }
        },
        "/api/me/change-password": {
            "post": {
                "tags": ["auth"],
                "summary": "修改密码（限流：5次/15分钟）",
                "requestBody": {
                    "required": true,
                    "content": { "application/json": { "schema": { "type": "object", "required": ["currentPassword", "newPassword"], "properties": { "currentPassword": { "type": "string" }, "newPassword": { "type": "string", "minLength": 6 } } } } }
                },
                "responses": { "200": { "description": "密码修改成功" }, "401": { "description": "当前密码不正确" } }
            }
        },
        "/api/me/api-settings/validate": {
            "post": {
                "tags": ["auth"],
                "summary": "验证 API 密钥有效性",
                "requestBody": {
                    "required": true,
                    "content": { "application/json": { "schema": { "type": "object", "required": ["provider", "credentials"], "properties": { "provider": { "type": "string" }, "credentials": { "type": "object" } } } } }
                },
                "responses": { "200": { "description": "验证结果" } }
            }
        },
        "/api/me/api-settings": {
            "get": { "tags": ["auth"], "summary": "获取 API 设置（脱敏）", "responses": { "200": { "description": "API 设置" } } },
            "put": {
                "tags": ["auth"],
                "summary": "更新 API 设置",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "provider": { "type": "string", "enum": ["tencent", "alibaba", "jimeng", "sdwebui"] },
                                    "credentials": { "type": "object" }
                                }
                            }
                        }
                    }
                },
                "responses": { "200": { "description": "设置成功" } }
            }
        },
        "/api/me/api-settings/{provider}": {
            "delete": {
                "tags": ["auth"],
                "summary": "删除 API 设置",
                "parameters": [{ "name": "provider", "in": "path", "required": true, "schema": { "type": "string" } }],
                "responses": { "200": { "description": "删除成功" } }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            }
        }
    },
    "security": [{ "cookieAuth": [] }]
}